(#11558) Add rspec-puppet coverage and cleaned up docs & code where necessary
I've added rspec-puppet coverage to some of the main classes in this commit.
The coverage is entirely complete as I still need to work on sunstone & econe
but since these classes need a good rework anyway I've left them for now.

And as I went through the code I've cleaned up parts and fixed documentation
where necessary. Primarily this entailed removing/adding/correcting parameter
docs or fixing params that shouldn't be there or are incorrect.

I've simplified the Rakefile and modified the spec_helper.rb to work with
rspec-puppet.